Park UpdateMotion (correct length, register-coloring tie); fix its item.h proto
Parked NON_MATCHING: 26 of 632 bytes, 13 instructions, and the function is the
correct length (whole-image diff == window diff), so everything downstream stays
byte-identical. Default ./Build check green, NON_MATCHING=UpdateMotion ./Build
compiles, symcheck clean.

The whole residual is one coloring decision in `min(model->n, motion->n)`: the
target coalesces the min result into $a0 with the `lh model->n` and keeps
`motion->n` live in $v1 (the taken branch is a free `move a0,v1`); our cc1 keeps
the `motion` POINTER in $a0, loads `motion->n` into a register the `slt` clobbers,
and must reload it in the branch. That $a0-vs-$a2 choice then dictates the
SetupSpline delay slot, the bone-fixup negu source, and the fixup stores. Not
reachable from C: field-vs-local, an explicit s16/s32 temp (adds a move + 4-byte
shift), reused `j`, both abs spellings -- all identical or worse; ~3400 permuter
iterations plus a 600s -j4 run never reached zero.

item.h: UpdateMotion's declaration `int id` -> `short id`. PSX.SYM's prototype is
`short mid`, and `short` is needed to reproduce the prologue sll/sra 16
sign-extension. Proven safe: all 5 matched callers pass either short-fitting
constants (0xf01/0xf03/0x810) or an already-short local, and ./Build check stays
byte-identical, so every recompiled caller is unchanged.

MotionManager fields this function's byte-correct accesses confirm: mid @0x00 s16,
count @0x02 s16, loop @0x04 s16, n @0x06 s16, model @0x0C (ModelArchiveType*),
motion @0x10 (MotionDataType*), motreg @0x14 (MotionRegistType*). Corroborating:
MotionDataType.n @0x0 u8, .sweep @0x1 u8; ModelArchiveType.n @0x64 s16, .object
@0x68; ModelType.rotate @0x50. All consistent with the existing item.h.

+++ src/main.exe/item.h
@@ -313,7 +313,7 @@ struct tag_TItem
 };                               /* sizeof = 0x58 (items[] stride) */
 
 extern void dispose_weapon_data_of_char_(Humanoid *h, int a);
-extern s16 UpdateMotion(MotionManager *m, int id);
+extern s16 UpdateMotion(MotionManager *m, short id);
 extern void MoveHumanoid(Humanoid *h, short a, short b);
 extern void UpdateCoordinate(ModelType *m);
 extern void DrawSprite(Sprite3D *s);
